+++ src/overlay/src/popup.js
@@ -17,17 +17,20 @@ KISSY.add('overlay/popup', function(S, Component, Overlay, undefined) {
     }
 
     Popup.ATTRS = {
-        trigger: {
+        trigger: {                          // 触发器
             setter:function(v) {
-                return S.one(v);
+                if (S.isString(v)) v = S.all(v);
+                return v;
             }
-        },                              // 触发器
-        triggerType: {value:'click'},   // 触发类型
+        },
+        triggerType: {value:'click'},       // 触发类型
+        currentTrigger: {
+        },
         mouseDelay: {
-            value: 100                  // triggerType 为 mouse 时, Popup 显示的延迟时间, 默认为 100ms
+            value: 100                      // triggerType 为 mouse 时, Popup 显示的延迟时间, 默认为 100ms
         },
 		toggle :{
-			value:false // triggerType 为 click 时, Popup 是否有toggle功能
+			value:false                     // triggerType 为 click 时, Popup 是否有toggle功能
 		}
     };
 
@@ -54,17 +57,18 @@ KISSY.add('overlay/popup', function(S, Component, Overlay, undefined) {
                 trigger = self.get("trigger"),
                 timer;
 
-            self.__mouseEnterPopup = function() {
+            self.__mouseEnterPopup = function(ev) {
                 self._clearHiddenTimer();
 
                 timer = S.later(function() {
-                    self.show();
+                    self._showing(ev);
                     timer = undefined;
                 }, self.get('mouseDelay'));
             };
 
-            trigger.on('mouseenter', self.__mouseEnterPopup);
-
+            S.each(trigger, function(el) {
+                S.one(el).on('mouseenter', self.__mouseEnterPopup);
+            });
 
             self._mouseLeavePopup = function() {
                 if (timer) {
@@ -75,7 +79,9 @@ KISSY.add('overlay/popup', function(S, Component, Overlay, undefined) {
                 self._setHiddenTimer();
             };
 
-            trigger.on('mouseleave', self._mouseLeavePopup);
+            S.each(trigger, function(el) {
+                S.one(el).on('mouseleave', self._mouseLeavePopup);
+            });
         },
 
         _bindContainerMouse: function() {
@@ -85,10 +91,10 @@ KISSY.add('overlay/popup', function(S, Component, Overlay, undefined) {
                 .on('mouseenter', self._clearHiddenTimer, self);
         },
 
-        _setHiddenTimer: function() {
+        _setHiddenTimer: function(ev) {
             var self = this;
             self._hiddenTimer = S.later(function() {
-                self.hide();
+                self._hiding(ev);
             }, self.get('mouseDelay'));
         },
 
@@ -102,21 +108,24 @@ KISSY.add('overlay/popup', function(S, Component, Overlay, undefined) {
 
         _bindTriggerClick: function() {
             var self = this;
-            self.__clickPopup = function(e) {
-                e.halt();
+            self.__clickPopup = function(ev) {
+                ev.halt();
 				if(self.get('toggle')){
-					if(self.get('el').css('visibility') == 'hidden'){
-						self.show();
-					}
-					else{
-						self.hide();
-					}
+                    self[self.get('visible')?'_hiding':'_showing'](ev);
 				}
-				else{
-					self.show();
-				}
+				else  self._showing(ev);
             };
-            self.get("trigger").on('click', self.__clickPopup);
+            S.each(self.get("trigger"), function(el) {
+                S.one(el).on('click', self.__clickPopup);
+            });
+        },
+        _showing: function(ev) {
+            this.set('currentTrigger', S.one(ev.target));
+            this.show();
+        },
+        _hiding: function(ev) {
+            this.set('currentTrigger', undefined);
+            this.hide();
         },
 
         destructor:function() {
@@ -124,14 +133,20 @@ KISSY.add('overlay/popup', function(S, Component, Overlay, undefined) {
             var t = self.get("trigger");
             if (t) {
                 if (self.__clickPopup) {
-                    t.detach('click', self.__clickPopup);
+                    S.each(t, function(el) {
+                        S.one(el).detach('click', self.__clickPopup);
+                    });
                 }
                 if (self.__mouseEnterPopup) {
-                    t.detach('mouseenter', self.__mouseEnterPopup);
+                    S.each(t, function(el) {
+                        S.one(el).detach('mouseenter', self.__mouseEnterPopup);
+                    });
                 }
 
                 if (self._mouseLeavePopup) {
-                    t.detach('mouseleave', self._mouseLeavePopup);
+                    S.each(t, function(el) {
+                        S.one(el).detach('mouseleave', self._mouseLeavePopup);
+                    });
                 }
             }
             if (self.get('el')) {